udacity/Data Science/Data Engineering with AWS/Data Warehouse/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blosc2==2.5.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.85
etils==1.7.0
flatbuffers==23.5.26
flax==0.8.1
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
idna==3.6
importlib_resources==6.1.2
jax==0.4.25
jaxlib==0.4.25
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.7
ndindex==1.8
nest-asyncio==1.6.0
numexpr==2.9.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.9
orbax-checkpoint==0.4.4
packaging==23.2
pandas==2.2.1
protobuf==4.25.3
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scipy==1.12.0
six==1.16.0
toolz==0.12.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.14.1
wurlitzer==3.0.3
zipp==3.17.0
